+++ src/Polysemy/Async.hs
@@ -10,11 +10,13 @@ module Polysemy.Async
 
     -- * Interpretations
   , asyncToIO
+  , asyncToIOFinal
   , lowerAsync
   ) where
 
 import qualified Control.Concurrent.Async as A
 import           Polysemy
+import           Polysemy.Final
 
 
 
@@ -32,12 +34,22 @@ data Async m a where
 makeSem ''Async
 
 ------------------------------------------------------------------------------
--- | A more flexible --- though less performant ---  version of 'lowerAsync'.
+-- | A more flexible --- though less performant ---
+-- version of 'asyncToIOFinal'.
 --
 -- This function is capable of running 'Async' effects anywhere within an
--- effect stack, without relying on an explicit function to lower it into 'IO'.
+-- effect stack, without relying on 'Final' to lower it into 'IO'.
 -- Notably, this means that 'Polysemy.State.State' effects will be consistent
 -- in the presence of 'Async'.
+--
+-- 'asyncToIO' is __unsafe__ if you're using 'await' inside higher-order actions
+-- of other effects interpreted after 'Async'.
+-- See <https://github.com/polysemy-research/polysemy/issues/205 Issue #205>.
+--
+-- Prefer 'asyncToIOFinal' unless you need to run pure, stateful interpreters
+-- after the interpreter for 'Async'.
+-- (Pure interpreters are interpreters that aren't expressed in terms of
+-- another effect or monad; for example, 'Polysemy.State.runState'.)
 --
 -- @since 1.0.0.0
 asyncToIO
@@ -57,9 +69,37 @@ asyncToIO m = withLowerToIO $ \lower _ -> lower $
     )  m
 {-# INLINE asyncToIO #-}
 
+------------------------------------------------------------------------------
+-- | Run an 'Async' effect in terms of 'A.async' through final 'IO'.
+--
+-- /Beware/: Effects that aren't interpreted in terms of 'IO'
+-- will have local state semantics in regards to 'Async' effects
+-- interpreted this way. See 'Final'.
+--
+-- Notably, unlike 'asyncToIO', this is not consistent with
+-- 'Polysemy.State.State' unless 'Polysemy.State.runStateIORef' is used.
+-- State that seems like it should be threaded globally throughout 'Async'
+-- /will not be./
+--
+-- Use 'asyncToIO' instead if you need to run
+-- pure, stateful interpreters after the interpreter for 'Async'.
+-- (Pure interpreters are interpreters that aren't expressed in terms of
+-- another effect or monad; for example, 'Polysemy.State.runState'.)
+--
+-- @since 1.2.0.0
+asyncToIOFinal :: Member (Final IO) r
+               => Sem (Async ': r) a
+               -> Sem r a
+asyncToIOFinal = interpretFinal $ \case
+  Async m -> do
+    ins <- getInspectorS
+    m'  <- runS m
+    liftS $ A.async (inspect ins <$> m')
+  Await a -> liftS (A.wait a)
+{-# INLINE asyncToIOFinal #-}
 
 ------------------------------------------------------------------------------
--- | Run an 'Async' effect via in terms of 'A.async'.
+-- | Run an 'Async' effect in terms of 'A.async'.
 --
 -- @since 1.0.0.0
 lowerAsync
@@ -80,4 +120,4 @@ lowerAsync lower m = interpretH
         Await a -> pureT =<< embed (A.wait a)
     )  m
 {-# INLINE lowerAsync #-}
-
+{-# DEPRECATED lowerAsync "Use 'asyncToIOFinal' instead" #-}