+++ src/Polysemy/Writer.hs
@@ -1,5 +1,4 @@
-{-# LANGUAGE TemplateHaskell #-}
-{-# LANGUAGE TupleSections   #-}
+{-# LANGUAGE TupleSections #-}
 
 module Polysemy.Writer
   ( -- * Effect
@@ -14,26 +13,27 @@ module Polysemy.Writer
     -- * Interpretations
   , runWriter
   , runWriterAssocR
+  , runWriterTVar
+  , writerToIOFinal
+  , writerToIOAssocRFinal
+  , writerToEndoWriter
 
     -- * Interpretations for Other Effects
   , outputToWriter
   ) where
 
+import Control.Concurrent.STM
+
 import Data.Bifunctor (first)
+import Data.Semigroup
 
 import Polysemy
 import Polysemy.Output
 import Polysemy.State
 
+import Polysemy.Internal.Writer
 
-------------------------------------------------------------------------------
--- | An effect capable of emitting and intercepting messages.
-data Writer o m a where
-  Tell   :: o -> Writer o m ()
-  Listen :: ∀ o m a. m a -> Writer o m (o, a)
-  Pass   :: m (o -> o, a) -> Writer o m a
 
-makeSem ''Writer
 
 ------------------------------------------------------------------------------
 -- | @since 0.7.0.0
@@ -94,31 +94,64 @@ runWriterAssocR
     => Sem (Writer o ': r) a
     -> Sem r (o, a)
 runWriterAssocR =
-  let
-    go :: forall o r a
-        . Monoid o
-       => Sem (Writer o ': r) a
-       -> Sem r (o -> o, a)
-    go =
-        runState id
-      . reinterpretH
-      (\case
-          Tell o -> do
-            modify' @(o -> o) (. (o <>)) >>= pureT
-          Listen m -> do
-            mm <- runT m
-            -- TODO(sandy): this is stupid
-            (oo, fa) <- raise $ go mm
-            modify' @(o -> o) (. oo)
-            pure $ fmap (oo mempty, ) fa
-          Pass m -> do
-            mm <- runT m
-            (o, t) <- raise $ runWriterAssocR mm
-            ins <- getInspectorT
-            let f = maybe id fst (inspect ins t)
-            modify' @(o -> o) (. (f o <>))
-            pure (fmap snd t)
-      )
-    {-# INLINE go #-}
-  in fmap (first ($ mempty)) . go
+    (fmap . first) (`appEndo` mempty)
+  . runWriter
+  . writerToEndoWriter
+  . raiseUnder
 {-# INLINE runWriterAssocR #-}
+
+--------------------------------------------------------------------
+-- | Transform a 'Writer' effect into atomic operations
+-- over a 'TVar' through final 'IO'.
+runWriterTVar :: (Monoid o, Member (Final IO) r)
+              => TVar o
+              -> Sem (Writer o ': r) a
+              -> Sem r a
+runWriterTVar tvar = runWriterSTMAction $ \o -> do
+  s <- readTVar tvar
+  writeTVar tvar $! s <> o
+{-# INLINE runWriterTVar #-}
+
+
+--------------------------------------------------------------------
+-- | Run a 'Writer' effect by transforming it into atomic operations
+-- through final 'IO'.
+--
+-- Internally, this simply creates a new 'TVar', passes it to
+-- 'runWriterTVar', and then returns the result and the final value
+-- of the 'TVar'.
+--
+-- /Beware/: Effects that aren't interpreted in terms of 'IO'
+-- will have local state semantics in regards to 'Writer' effects
+-- interpreted this way. See 'Final'.
+writerToIOFinal :: (Monoid o, Member (Final IO) r)
+                => Sem (Writer o ': r) a
+                -> Sem r (o, a)
+writerToIOFinal sem = do
+  tvar <- embedFinal $ newTVarIO mempty
+  res  <- runWriterTVar tvar sem
+  end  <- embedFinal $ readTVarIO tvar
+  return (end, res)
+{-# INLINE writerToIOFinal #-}
+
+--------------------------------------------------------------------
+-- | Like 'writerToIOFinal'. but right-associates uses of '<>'.
+--
+-- This asymptotically improves performance if the time complexity of '<>'
+-- for the 'Monoid' depends only on the size of the first argument.
+--
+-- You should always use this instead of 'writerToIOFinal' if the monoid
+-- is a list, such as 'String'.
+--
+-- /Beware/: Effects that aren't interpreted in terms of 'IO'
+-- will have local state semantics in regards to 'Writer' effects
+-- interpreted this way. See 'Final'.
+writerToIOAssocRFinal :: (Monoid o, Member (Final IO) r)
+                      => Sem (Writer o ': r) a
+                      -> Sem r (o, a)
+writerToIOAssocRFinal =
+    (fmap . first) (`appEndo` mempty)
+  . writerToIOFinal
+  . writerToEndoWriter
+  . raiseUnder
+{-# INLINE writerToIOAssocRFinal #-}